update UI for app
